# docs/api/api-docs.v1.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"payment","description":"결제 관련 API"},{"name":"coupon","description":"쿠폰 관련 API"},{"name":"order","description":"주문 관련 API"},{"name":"product","description":"상품 관련 API"},{"name":"user","description":"유저 관련 API"}],"paths":{"/payments":{"post":{"tags":["payment"],"summary":"결제 요청","description":"유저는 취소되지 않은 주문에 대해 포인트를 사용하여 결제를 요청할 수 있습니다.","operationId":"requestPayment","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/orders":{"get":{"tags":["OrderController","order"],"summary":"전체 주문 조회","description":"유저는 모든 주문의 내용을 조회합니다.","operationId":"getAllOrders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["OrderController","order"],"summary":"상품 주문","description":"유저는 아직 재고가 남아있는 상품을 주문합니다.","operationId":"createOrder","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/orders/{orderId}/cancel":{"post":{"tags":["OrderController","order"],"summary":"주문 취소","description":"유저는 [결제전, 결제 완료, 배송전] 상태의 주문을 취소할 수 있습니다.","operationId":"cancelOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/coupons/{couponId}/issue":{"post":{"tags":["coupon","CouponController"],"summary":"선착순 쿠폰 발급","description":"유저는 선착순으로 제공되는 쿠폰을 발급 받아 등록합니다.","operationId":"issueCoupon","parameters":[{"name":"couponId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/Users/{userId}/pointCharge":{"post":{"tags":["user"],"summary":"포인트 충전","description":"유저는 상품 구매를 위한 포인트를 충전합니다","operationId":"chargeUserPoint","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/products":{"get":{"tags":["product"],"summary":"상품 조회","description":"모든 상품을 조회합니다.","operationId":"getAllProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/products/{productId}":{"get":{"tags":["product"],"summary":"상품 조회","description":"특정 상품을 조회합니다.","operationId":"getProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/products/products/popular":{"get":{"tags":["product"],"summary":"인기 판매 상품 조회","description":"현재 일을 기준으로 최근 3일간 가장 많이 판매된 상품 N개를 조회합니다.","operationId":"getPopularProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/payments/{paymentId}":{"get":{"tags":["payment"],"summary":"결제 조회","description":"유저는 주문의 결제 상태를 조회할 수 있습니다.","operationId":"getPaymentById","parameters":[{"name":"paymentId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/orders/{orderId}":{"get":{"tags":["OrderController","order"],"summary":"주문 조회","description":"유저는 단건 주문의 내용을 조회합니다.","operationId":"getOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/coupons":{"get":{"tags":["coupon","CouponController"],"summary":"쿠폰 조회","description":"유저가 현재 사용할 수 있는 쿠폰을 조회합니다.","operationId":"getAllCoupons","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/Users/{userId}/point":{"get":{"tags":["user"],"summary":"포인트 조회","description":"유저는 자신이 보유한 포인트를 조회합니다","operationId":"getUserPoints","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{}}